Fixing memory leaks in integrateSD (incomplete).

+++ stocUCED/src/SUC_master.cpp
@@ -279,7 +279,7 @@ void SUCmaster::preprocessing ()
 				}
 				else {
 					it = inst->simulations.mapVarNamesToIndex.find(genPtr->name);
-					for (int s=0; s<inst->simulations.vals.size(); s++) {
+					for (int s=0; s < (int) inst->simulations.vals.size(); s++) {
 						expCapacity[g][t] += 1.0/(double)(inst->simulations.vals.size()) * min( inst->simulations.vals[s][period][it->second], genPtr->maxCapacity );
 					}
 				}
@@ -691,7 +691,7 @@ void SUCmaster::formulate (instance &inst, ProblemType probType, ModelType model
 				IloExpr expr (env);
 				
 				// production (iterate over connected generators)
-				for (int g=0; g<busPtr->connectedGenerators.size(); g++) {
+				for (int g = 0; g < (int) busPtr->connectedGenerators.size(); g++) {
 					expr += p[ busPtr->connectedGenerators[g]->id ][t];
 				}
 				
@@ -921,7 +921,7 @@ bool SUCmaster::getGenState(int genId, int period) {
 	if (reqSolnComp < 0) {										// all generators are assumed to be online, for a long time, at t=0.
 		return true;
 	}
-	else if (reqSolnComp < inst->solution.x[genId].size()) {	// return the corresponding solution
+	else if (reqSolnComp < (int) inst->solution.x[genId].size()) {	// return the corresponding solution
 		return round(inst->solution.x[genId][reqSolnComp]);
 	}
 	else {														// asking what's beyond the planning horizon, we return the last solution
@@ -938,7 +938,7 @@ void SUCmaster::setGenState(int genId, int period, double value) {
 	int solnComp = beginMin/runParam.ED_resolution + period*numBaseTimePerPeriod;
 	
 	// set the solution
-	if (solnComp >= 0 && solnComp < inst->solution.x[genId].size()) {
+	if (solnComp >= 0 && solnComp < (int) inst->solution.x[genId].size()) {
 		for (int t=solnComp; t<solnComp+numBaseTimePerPeriod; t++) {
 			inst->solution.x[genId][t] = value;
 		}
@@ -970,7 +970,7 @@ double SUCmaster::getEDGenProd(int genId, int period) {
 		cout << "Error: Initial production levels are not available" << endl;
 		exit(1);
 	}
-	else if (reqSolnComp < inst->solution.x[genId].size()) {	// return the corresponding solution
+	else if (reqSolnComp < (int) inst->solution.x[genId].size()) {	// return the corresponding solution
 		return inst->solution.g_ED[genId][reqSolnComp];
 	}
 	else {														// asking what's beyond the planning horizon, we return the last solution
@@ -988,7 +988,7 @@ void SUCmaster::setGenProd(int genId, int period, double value) {
 	int solnComp = beginMin/runParam.ED_resolution + period*numBaseTimePerPeriod;
 	
 	// set the solution
-	if (solnComp >= 0 && solnComp < inst->solution.g_UC[genId].size()) {
+	if (solnComp >= 0 && solnComp < (int) inst->solution.g_UC[genId].size()) {
 		for (int t=solnComp; t<solnComp+numBaseTimePerPeriod; t++) {
 			inst->solution.g_UC[genId][t] = value;
 		}
